+++ src/generators/audio_stream_osc.h
@@ -42,7 +42,7 @@ public:
 	virtual int32_t _get_loop_count() const override;
 	virtual double _get_playback_position() const override;
 	virtual void _seek(double p_time) override;
-	virtual int32_t _mix(AudioFrame *p_buffer, double p_rate_scale, int32_t p_frames) override;
+	virtual int _mix(AudioFrame *p_buffer, float p_rate_scale, int p_frames) override;
 	virtual void _tag_used_streams() override;
 };
